refactoring renderElement
