+++ src/view/abstract-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render';
+import { createElement } from '../utils/render';
 
 class AbstractView {
   #element = null;
